tests: convert util test to a unit test
Merge tests/misc/util with existing util unit test.
