--- tests/misc/util/prj.conf
@@ -1,1 +1,0 @@
-CONFIG_ZTEST=y
