--- tests/misc/util/testcase.yaml
@@ -1,3 +1,0 @@
-tests:
-  misc.util:
-    tags: util
